Fix int type benchmark_random_interleaving_gtest for macos in fr-1051
Committer: Hai Huang <haih@google.com>

On branch fr-1051
Your branch is up to date with 'origin/fr-1051'.

Changes to be committed:
modified:   test/benchmark_random_interleaving_gtest.cc

+++ test/benchmark_random_interleaving_gtest.cc
@@ -68,7 +68,7 @@ static void BM_Match1(benchmark::State& state) {
   const int64_t arg = state.range(0);
 
   for (auto _ : state) {}
-  queue->Put(StrFormat("BM_Match1/%ld", arg));
+  queue->Put(StrFormat("BM_Match1/%d", static_cast<int>(arg)));
 }
 BENCHMARK(BM_Match1)
     ->Iterations(100)
@@ -83,7 +83,7 @@ static void BM_MatchOverhead(benchmark::State& state) {
   const int64_t arg = state.range(0);
 
   for (auto _ : state) {}
-  queue->Put(StrFormat("BM_MatchOverhead/%ld", arg));
+  queue->Put(StrFormat("BM_MatchOverhead/%d", static_cast<int>(arg)));
 }
 BENCHMARK(BM_MatchOverhead)
     ->Iterations(100)
